requirements.txt
numpy==1.21.0
imageio==2.9.0
opencv-python
colour-science==0.3.16
colour-demosaicing==0.1.7